refactor: generalize validator functions
