+++ src/constants/command.js
@@ -1,34 +1,37 @@
-import { validateAdd } from '../validators/validate-add.js';
-import { validateRn } from '../validators/validate-rn.js';
-import { validateCp } from '../validators/validate-cp.js';
-import { validateMv } from '../validators/validate-mv.js';
-import { validateRm } from '../validators/validate-rm.js';
-import { validateCat } from '../validators/validate-cat.js';
-import { validateCompress } from '../validators/validate-compress.js';
-import { validateDecompress } from '../validators/validate-decompress.js';
-import { validateHash } from '../validators/validate-hash.js';
 import { up } from '../commands/nwd/up.js';
 import { ls } from '../commands/nwd/ls.js';
-import { validateCd } from '../validators/validate-cd.js';
 import { validateOs } from '../validators/validate-os.js';
+
+import { validateCommandWithTwoArgs } from '../validators/validate-command-with-two-args.js';
+import { cp } from '../commands/file/cp.js';
+import { mv } from '../commands/file/mv.js';
+import { rn } from '../commands/file/rn.js';
+import { compress } from '../commands/brotli/compress.js';
+import { decompress } from '../commands/brotli/decompress.js';
+import { validateCommandWithOneArg } from '../validators/validate-command-with-one-arg.js';
+import { cat } from '../commands/file/cat.js';
+import { add } from '../commands/file/add.js';
+import { rm } from '../commands/file/rm.js';
+import { hash } from '../commands/hash.js';
+import { cd } from '../commands/nwd/cd.js';
 
 import {SPACE} from './space.js';
 
 const commands = {
     UP: {name: 'up', func: up},
-    CD: {name: `cd${SPACE}`, func: validateCd},
+    CD: {name: `cd${SPACE}`, func: (input) => validateCommandWithOneArg(input, cd, 'cd ')},
     LS: {name: 'ls', func: ls},
     
-    CAT: {name: `cat${SPACE}`, func: validateCat},
-    ADD: {name: `add${SPACE}`, func: validateAdd},
-    RN: {name: `rn${SPACE}`, func: validateRn},
-    CP: {name: `cp${SPACE}`, func: validateCp},
-    MV: {name: `mv${SPACE}`, func: validateMv},
-    RM: {name: `rm${SPACE}`, func: validateRm},
+    CAT: {name: `cat${SPACE}`, func: (input) => validateCommandWithOneArg(input, cat, 'cat ')},
+    ADD: {name: `add${SPACE}`, func: (input) => validateCommandWithOneArg(input, add, 'add ')},
+    RN: {name: `rn${SPACE}`, func: (input) => validateCommandWithTwoArgs(input, rn, 'rn')},
+    CP: {name: `cp${SPACE}`, func: (input) => validateCommandWithTwoArgs(input, cp, 'cp')},
+    MV: {name: `mv${SPACE}`, func: (input) => validateCommandWithTwoArgs(input, mv, 'rn')},
+    RM: {name: `rm${SPACE}`, func: (input) => validateCommandWithOneArg(input, rm, 'rm ')},
 
-    COMPRESS: {name: `compress${SPACE}`, func: validateCompress},
-    DECOMPRESS: {name: `decompress${SPACE}`, func: validateDecompress},
-    HASH: {name: `hash${SPACE}`, func: validateHash},
+    COMPRESS: {name: `compress${SPACE}`, func: (input) => validateCommandWithTwoArgs(input, compress, 'compress')},
+    DECOMPRESS: {name: `decompress${SPACE}`, func: (input) => validateCommandWithTwoArgs(input, decompress, 'decompress')},
+    HASH: {name: `hash${SPACE}`, func: (input) => validateCommandWithOneArg(input, hash, 'hash ')},
 
     OS: {name: `os `, func: validateOs}
 };
