--- src/validators/validate-command-with-one-args.js
@@ -1,5 +1,0 @@
-add
-cd
-cat
-hash
-rm
